Code Formatting

+++ src/main/java/com/vinodh/repository/AuthenticationUserDetailsDAO.java
@@ -22,49 +22,40 @@ import com.vinodh.domain.AuthenticationUser;
  *
  */
 @Repository
-public interface AuthenticationUserDetailsDAO { 
- 
-	 final String LIST_ALL_USERS = "SELECT USER_ID,"
-											+ "USER_FIRST_NAME,"
-											+ "USER_LAST_NAME,"
-											+ "USER_EMAIL,"
-											+ "USER_PHONE,"
-											+ "USER_ROLE,"
-											+ "APPLICATION_NAME,"
-											+ "LAST_MODIFIED_BY,"
-											+ "LAST_MODIFIED_DATE"
-											+ " FROM VINODH.AUTHENTICATION_USER_DETAILS";
-							 
-	 final String GET_USER = "SELECT * FROM VINODH.AUTHENTICATION_USER_DETAILS WHERE USER_ID = #{userId}";
-	 
-	 final String INSERT_USER ="INSERT INTO	VINODH.AUTHENTICATION_USER_DETAILS "
-	 						+ "(USER_ID,USER_FIRST_NAME,USER_LAST_NAME,USER_EMAIL,USER_PHONE,"
-	 						+ "USER_ROLE,APPLICATION_NAME,LAST_MODIFIED_BY,LAST_MODIFIED_DATE)"
-	 						+ " VALUES(#{authenticationUser.userId},#{authenticationUser.userFirstName},#{authenticationUser.userLastName},"
-	 						+ "#{authenticationUser.userEmail},#{authenticationUser.userPhone},"
-	 						+ "#{authenticationUser.role},#{authenticationUser.applicationName},#{loggedInUser},SYSDATE)";
-	 
-	 final String UPDATE_USER = "UPDATE VINODH.AUTHENTICATION_USER_DETAILS"
-			 					+ " SET USER_ROLE = #{authenticationUser.role},"
-			 					+ " LAST_MODIFIED_BY = 	#{loggedInUser}"
-			 					+ " WHERE USER_ID = #{authenticationUser.userId}";
-	 
-	 final String MERGE_USER = "MERGE INTO VINODH.AUTHENTICATION_USER_DETAILS SOURCE"
-			 				+ " USING"
-			 				+ " (SELECT 1 FROM DUAL) TARGET" /* Merge operation on Same Table, so using the DUAL*/
-			 				+ "	ON (SOURCE.USER_ID = #{authenticationUser.userId})"
-			 				+ " WHEN MATCHED THEN " /* Update if already exists*/
-			 				+ " UPDATE SET SOURCE.USER_ROLE = #{authenticationUser.role}, SOURCE.LAST_MODIFIED_BY = #{loggedInUser}"
-		 					+ " WHEN NOT MATCHED THEN " /* Insert a New Record if User doesn't exists already*/
-		 					+ " INSERT (USER_ID,USER_FIRST_NAME,USER_LAST_NAME,USER_EMAIL,USER_PHONE,USER_ROLE,APPLICATION_NAME,LAST_MODIFIED_BY,LAST_MODIFIED_DATE)"
-		 					+ " VALUES(#{authenticationUser.userId},#{authenticationUser.userFirstName},#{authenticationUser.userLastName},#{authenticationUser.userEmail},#{authenticationUser.userPhone},#{authenticationUser.role},#{authenticationUser.applicationName},#{loggedInUser},SYSDATE)";
-	 
-	 final String DELETE_USER = "DELETE FROM VINODH.AUTHENTICATION_USER_DETAILS"
-	 							+ " WHERE USER_ID = #{userId}";
-	 
+public interface AuthenticationUserDetailsDAO {
+
+	final String LIST_ALL_USERS = "SELECT * FROM VINODH.AUTHENTICATION_USER_DETAILS";
+
+	final String GET_USER = "SELECT * FROM VINODH.AUTHENTICATION_USER_DETAILS WHERE USER_ID = #{userId}";
+
+	final String INSERT_USER = "INSERT INTO	VINODH.AUTHENTICATION_USER_DETAILS "
+			+ "(USER_ID,USER_FIRST_NAME,USER_LAST_NAME,USER_EMAIL,USER_PHONE,"
+			+ "USER_ROLE,APPLICATION_NAME,LAST_MODIFIED_BY,LAST_MODIFIED_DATE)"
+			+ " VALUES(#{authenticationUser.userId},#{authenticationUser.userFirstName},#{authenticationUser.userLastName},"
+			+ "#{authenticationUser.userEmail},#{authenticationUser.userPhone},"
+			+ "#{authenticationUser.role},#{authenticationUser.applicationName},#{loggedInUser},SYSDATE)";
+
+	final String UPDATE_USER = "UPDATE VINODH.AUTHENTICATION_USER_DETAILS"
+			+ " SET USER_ROLE = #{authenticationUser.role}, LAST_MODIFIED_BY = 	#{loggedInUser}"
+			+ " WHERE USER_ID = #{authenticationUser.userId}";
+
+	final String MERGE_USER = "MERGE INTO VINODH.AUTHENTICATION_USER_DETAILS SOURCE USING"
+			+ " (SELECT 1 FROM DUAL) TARGET"
+			/* Merge operation on Same Table, so using the DUAL */
+			+ "	ON (SOURCE.USER_ID = #{authenticationUser.userId})"
+			/* Update if already exists */
+			+ " WHEN MATCHED THEN "
+			+ " UPDATE SET SOURCE.USER_ROLE = #{authenticationUser.role}, SOURCE.LAST_MODIFIED_BY = #{loggedInUser}"
+			/* Insert a New Record if User doesn't exists already */
+			+ " WHEN NOT MATCHED THEN "
+			+ " INSERT (USER_ID,USER_FIRST_NAME,USER_LAST_NAME,USER_EMAIL,USER_PHONE,USER_ROLE,APPLICATION_NAME,LAST_MODIFIED_BY,LAST_MODIFIED_DATE)"
+			+ " VALUES(#{authenticationUser.userId},#{authenticationUser.userFirstName},#{authenticationUser.userLastName},#{authenticationUser.userEmail},#{authenticationUser.userPhone},#{authenticationUser.role},#{authenticationUser.applicationName},#{loggedInUser},SYSDATE)";
+
+	final String DELETE_USER = "DELETE FROM VINODH.AUTHENTICATION_USER_DETAILS WHERE USER_ID = #{userId}";
+
 	@Select(LIST_ALL_USERS)
 	@Results(value = { @Result(property = "role", column = "USER_ROLE", javaType = String.class) })
-	
+
 	List<AuthenticationUser> listAllAuthUsers();
 
 	@Select(GET_USER)
